fix(presence): CI breakages from the presence_mode commit
Four issues surfaced by CI on the feature commit:

1. **Backend test DB migration failure** — auth.permissions has NOT NULL
   resource + action columns (see 001000005_auth_permissions.go). My
   migration only set name + description, so the INSERT violated the
   NOT NULL constraint and the test DB failed to initialise. Fixed by
   including resource='users', action='checkin', is_system=TRUE in the
   INSERT. The unique index on (resource, action) is another legitimate
   conflict path alongside the name index; ON CONFLICT (name) still
   covers the re-run case because names are globally unique.

2. **Backend staticcheck ST1005** — the German error constant ended with
   a period. Switched the trailing full stop for an em-dash-joined clause
   so staticcheck is happy and the frontend copy still reads correctly.

3. **Frontend sidebar + page tests (239 failures)** — every test that
   rendered a component reading presence state failed with "No
   TenantContext export on the tenant-provider mock". The global mock in
   src/test/setup.ts doesn't replicate createContext output, and my
   hook was reaching into TenantContext directly. Fixed by co-locating
   usePresenceMode inside tenant-provider.tsx (where the context now
   stays unexported) and adding the hook to the global test mock so
   every component that calls it works out of the box.

4. **Knip "unused exports"** — TenantContext was exported so the hook in
   ~/lib/hooks could subscribe; knip flagged it as unused because its
   static analysis doesn't follow the ~/-alias import through the hook
   file. Co-locating the hook (see #3) removes the export entirely and
   eliminates the false positive. Also deleted the now-empty hook shim
   at ~/lib/hooks/use-presence-mode.ts and migrated its 3 tests into
   tenant-provider.test.tsx alongside useTenant / useTenantSlugSafe.

Consumers (sidebar, binary-mode-guard, student-presence-badge) now
import usePresenceMode from "~/components/tenant/tenant-provider"
directly, matching how useTenant and useTenantSlugSafe are imported.

Verified:
- cd backend && go vet ./... && gofmt -l . — clean
- cd frontend && pnpm run typecheck — only pre-existing Sentry errors
- cd frontend && pnpm run lint — 0 errors
- cd frontend && pnpm knip — no Unused exports
- cd frontend && pnpm vitest run on the six changed test files —
  125 pass (was 120; gained 3 from the hook-test migration +
  2 repaired paths)
