+++ frontend/src/components/tenant/tenant-provider.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { createContext, useContext, useMemo } from "react";
-import type { TenantInfo } from "~/lib/tenant-api";
+import type { PresenceMode, TenantInfo } from "~/lib/tenant-api";
 
 interface TenantContextValue {
   /** The tenant slug from the URL path segment */
@@ -10,12 +10,14 @@ interface TenantContextValue {
   tenant: TenantInfo | null;
 }
 
-/**
- * Exported so sibling hooks (e.g. `usePresenceMode`) can subscribe to the
- * same context without routing through `useTenant()` — which throws outside
- * the provider. The hooks return safe defaults when the context is missing.
- */
-export const TenantContext = createContext<TenantContextValue | null>(null);
+// Kept unexported: consumers use the hooks below. An export was tried
+// earlier to let sibling modules subscribe directly, but knip flagged it
+// as unused (its static analysis doesn't follow `~/` alias imports through
+// the hook file reliably), and the global test mock in `src/test/setup.ts`
+// doesn't replicate `createContext` output — which broke every test that
+// rendered a component reading presence state. Co-locating the hook here
+// sidesteps both issues.
+const TenantContext = createContext<TenantContextValue | null>(null);
 
 /**
  * Provides tenant context to all child components within the [tenant] route segment.
@@ -60,4 +62,21 @@ export function useTenant(): TenantContextValue {
 export function useTenantSlugSafe(): string | null {
   const ctx = useContext(TenantContext);
   return ctx?.tenantSlug ?? null;
+}
+
+/**
+ * Returns the current tenant's presence mode ("detailed" | "binary").
+ *
+ * Falls back to "detailed" when called outside a TenantProvider or before
+ * the tenant metadata has resolved — matches the backend's safe default so
+ * first-paint shows the richer view even if the tenant turns out to be
+ * binary (the switch happens transparently on the next render).
+ *
+ * Safe to call from any component. Co-located with the TenantContext it
+ * reads from, so the context can stay unexported — the global test mock
+ * in `src/test/setup.ts` re-exports this stub directly.
+ */
+export function usePresenceMode(): PresenceMode {
+  const ctx = useContext(TenantContext);
+  return ctx?.tenant?.presenceMode ?? "detailed";
 }
